\name{labbonus-package}
\alias{labbonus-package}
\alias{labbonus}
\docType{package}
\title{
\packageTitle{labbonus}
}
\description{
\packageDescription{labbonus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{labbonus}
\packageIndices{labbonus}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{labbonus}

Maintainer: \packageMaintainer{labbonus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
