% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labbonus.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{print out formula, data and Coefficients of current linear regression.}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{ridgereg object}

\item{...}{other objects}
}
\description{
print out formula, data and Coefficients of current linear regression.
}
