% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labbonus.R
\name{ridgereg}
\alias{ridgereg}
\title{A S3 class provide ridgereg which provide same functions like linear regression lm().}
\usage{
ridgereg(formula, data, lambda = 0, qr = FALSE)
}
\arguments{
\item{formula}{Formula}

\item{data}{Dataframe}

\item{lambda}{Lambda}

\item{qr}{use qr decomposition or not}
}
\value{
ridgereg object(list) contain all the calculated values regarding the linear regression
}
\description{
A S3 class provide ridgereg which provide same functions like linear regression lm().
}
