% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labbouns-package.R
\docType{package}
\name{labbonus-package}
\alias{labbonus-package}
\alias{labbonus}
\title{Lab Bonus of Advanced Programming in R}
\description{
This package include following S3 class\cr
\cr 
1. ridgereg - a class provide linear regression  that is similar as lm()\cr 
\cr
}
