% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Delay & Distance Relationship}
\usage{
visualize_airport_delays()
}
\value{
Return plot
}
\description{
visualize_airport_delays() creates a plot that visualize the relation between
mean(distance) and mean(delay) using ggplot2 library(using same destination)
}
